+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 ```
+* Commit: 50959a7 - v10
+| Author: FJrodafo - Sun, 6 Oct 2024 19:35:14 +0100
+| 
 * Commit: 48c11e0 - v9
 | Author: FJrodafo - Mon, 26 Aug 2024 20:01:36 +0100
 |